feat(vm-debug): extract generic DebugHooks + DebugServer from twig-vm (VMDEBUG01)
New vm-debug crate lifts the generic debug substrate out of twig-vm so
per-language DAP adapters (basic-dap, nib-dap, oct-dap — task #37) can
depend on a single substrate without pulling in twig-vm's full
Twig→IIR→Lispy stack.

## What's in vm-debug

- DebugFrame trait — abstracts FrameView's per-VM frame access
  (register_names, read_register)
- DebugHooks trait — before_instruction(fn_name, depth, pc, &dyn DebugFrame)
- StopReason enum — Breakpoint / Step / Pause / Entry (wire format)
- MAX_LINE_BYTES + read_line_capped — DoS guard
- DebugServer — TCP-backed production DebugHooks impl with stable
  per-fn slot indexing via new_with_module
- 12 unit tests, all passing

## twig-vm changes (source-compatible)

- twig_vm::debug::FrameView still exists; now impls vm_debug::DebugFrame
- twig_vm::debug::DebugHooks re-exports vm_debug::DebugHooks
- twig_vm::debug_server::{DebugServer, StopReason, MAX_LINE_BYTES}
  re-export from vm_debug
- Dispatch loop unchanged — &view auto-coerces to &dyn DebugFrame
  through the new FrameView impl

External code that wrote `twig_vm::debug::DebugHooks` /
`twig_vm::debug_server::DebugServer` continues to compile unchanged.

## Versions

- vm-debug 0.1.0 (new)
- twig-vm 0.22.0 -> 0.23.0

## Tests

- 12 new vm-debug tests pass
- All 171 existing twig-vm tests pass
- All 2 twig-dap end-to-end tests pass
- All 17 lang-aot smoke tests pass

Task #36 in the multi-language debugger track.  Next: task #37
({basic,nib,oct}-dap crates implementing LanguageDebugAdapter).
