+++ code/packages/rust/twig-vm/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "twig-vm"
-version = "0.22.0"
+version = "0.23.0"
 edition = "2021"
-description = "LANG47 + LANG52 + LANG55 + LANG56 + LANG57 + LANG62 + LANG64 + LANG66 — string heap objects, builtins, host I/O, higher-order list ops, multi-file entry points, record/union/match runtime smoke tests; MAX_DISPATCH_DEPTH 256→4096 (LANG62), 4096→65536 (LANG64), 65536→131072 (LANG66) for self-hosted compiler pipeline.  Path A increment 6a: exec_const dispatches `const 0 [ref<LispyPair>]` to LispyValue::NIL.  Path A increment 6b: exec_alloc + exec_field_store for typed cons cells."
+description = "LANG47 + LANG52 + LANG55 + LANG56 + LANG57 + LANG62 + LANG64 + LANG66 — string heap objects, builtins, host I/O, higher-order list ops, multi-file entry points, record/union/match runtime smoke tests; MAX_DISPATCH_DEPTH 256→4096 (LANG62), 4096→65536 (LANG64), 65536→131072 (LANG66) for self-hosted compiler pipeline.  Path A increment 6a: exec_const dispatches `const 0 [ref<LispyPair>]` to LispyValue::NIL.  Path A increment 6b: exec_alloc + exec_field_store for typed cons cells.  VMDEBUG01: generic DebugHooks + DebugServer extracted to the new vm-debug crate (source-compatible — twig-vm re-exports everything)."
 
 [dependencies]
 twig-ir-compiler   = { path = "../twig-ir-compiler" }
@@ -12,6 +12,9 @@ twig-module-driver = { path = "../twig-module-driver" }
 lispy-runtime      = { path = "../lispy-runtime" }
 lang-runtime-core  = { path = "../lang-runtime-core" }
 interpreter-ir     = { path = "../interpreter-ir" }
+# VMDEBUG01: generic DebugHooks + DebugServer extracted from this crate
+# into vm-debug so basic/nib/oct-dap can depend on a single substrate.
+vm-debug           = { path = "../vm-debug" }
 serde_json         = "1"
 
 [[bin]]
