+++ code/packages/rust/twig-vm/src/debug.rs
@@ -1,23 +1,32 @@
-//! Debug-mode hooks for the dispatch loop.
+//! Debug-mode glue between the twig-vm dispatcher and the generic
+//! [`vm_debug`] substrate.
 //!
-//! When a [`DebugHooks`] implementation is supplied to
-//! [`crate::dispatch::run_with_debug`], the dispatcher calls
-//! [`DebugHooks::before_instruction`] **between every instruction** at
-//! every recursion depth.  The hook owns whatever blocking-on-TCP logic
-//! is needed to support pause / step / breakpoint semantics — the
-//! dispatcher itself stays a tight loop over IIR instructions.
+//! ## What lives where (post-VMDEBUG01)
+//!
+//! - **`vm_debug` crate**: the generic [`DebugHooks`] trait, the
+//!   [`DebugFrame`] trait, the [`StopReason`] enum, and the
+//!   production TCP-backed [`DebugServer`].  Shared across every
+//!   LANG-VM-based interpreter.
+//! - **This module**: the concrete [`FrameView`] (a borrow of
+//!   `dispatch::Frame`) plus its `impl vm_debug::DebugFrame`.  Keeps
+//!   the link between twig-vm's private `Frame` type and the public
+//!   debug surface in one place.
+//!
+//! For back-compat, this module also re-exports
+//! [`vm_debug::DebugHooks`] so callers that wrote
+//! `twig_vm::debug::DebugHooks` keep working.
 //!
 //! ## Why a trait, not a concrete type
 //!
-//! The trait keeps `twig-vm` free of TCP / DAP concerns: those live in
-//! [`crate::debug_server`] (the production hook) and in tests (mock
-//! hooks that just count calls or assert on locations).  Three callers,
-//! one signature.
+//! The trait keeps `twig-vm` free of TCP / DAP concerns: those live
+//! in [`vm_debug::DebugServer`] (the production hook) and in tests
+//! (mock hooks that just count calls or assert on locations).  Three
+//! callers, one signature.
 //!
 //! ## Call-stack reconstruction
 //!
-//! The hook receives `(fn_name, depth, pc, frame)` on every call and can
-//! reconstruct the live call stack by tracking depth deltas:
+//! The hook receives `(fn_name, depth, pc, frame)` on every call and
+//! can reconstruct the live call stack by tracking depth deltas:
 //!
 //! ```text
 //! before_instruction(main, depth=0, pc=0)  → stack = [main:0]
@@ -27,11 +36,15 @@
 //! ```
 //!
 //! `main:?` between the push and the pop is filled in by the most-
-//! recently-reported pc for that depth.  Holding one entry per depth in
-//! a `Vec<(String, usize)>` is a single 16-byte update per safepoint —
-//! cheap.
+//! recently-reported pc for that depth.  Holding one entry per depth
+//! in a `Vec<(String, usize)>` is a single 16-byte update per
+//! safepoint — cheap.
 
 use crate::dispatch::Frame;
+
+// Re-export so existing callers (`use twig_vm::debug::DebugHooks;`)
+// keep working without changes.
+pub use vm_debug::{DebugFrame, DebugHooks};
 
 // ---------------------------------------------------------------------------
 // FrameView — read-only snapshot of one Frame
@@ -39,10 +52,14 @@ use crate::dispatch::Frame;
 
 /// Read-only view into a [`Frame`] for the debug hook.
 ///
-/// The underlying `Frame` is private to `dispatch`.  `FrameView` exposes
-/// just the bits a debugger needs (variable lookup by name, name
-/// enumeration) without leaking `Frame` itself or the
+/// The underlying `Frame` is private to `dispatch`.  `FrameView`
+/// exposes just the bits a debugger needs (variable lookup by name,
+/// name enumeration) without leaking `Frame` itself or the
 /// register-storage details.
+///
+/// Implements [`vm_debug::DebugFrame`] so it can be passed to any
+/// hook the generic [`DebugServer`](vm_debug::DebugServer) plugs
+/// into.
 pub struct FrameView<'a> {
     inner: &'a Frame,
 }
@@ -63,8 +80,8 @@ impl<'a> FrameView<'a> {
         self.inner.register_names()
     }
 
-    /// Return a printable representation of `name`'s current value, or
-    /// `None` if the register is not bound in this frame.
+    /// Return a printable representation of `name`'s current value,
+    /// or `None` if the register is not bound in this frame.
     ///
     /// We return `String` rather than `LispyValue` to keep the
     /// debug-server / DAP surface free of `lispy-runtime` types.
@@ -73,48 +90,13 @@ impl<'a> FrameView<'a> {
     }
 }
 
-// ---------------------------------------------------------------------------
-// DebugHooks trait
-// ---------------------------------------------------------------------------
-
-/// Sole interface between [`crate::dispatch`] and any debug agent.
-///
-/// `dispatch` calls [`Self::before_instruction`] right before executing
-/// each IIR instruction.  Implementations are free to:
-///
-/// 1. Process incoming wire-protocol commands (set/clear breakpoint).
-/// 2. Detect that the current `(fn_name, pc)` is a breakpoint and emit
-///    a `stopped` event.
-/// 3. **Block** on the wire (e.g. waiting for `continue` after a stop)
-///    — the dispatcher resumes only when the hook returns.
-/// 4. Implement single-step by remembering "stop on the *next*
-///    instruction" between calls.
-///
-/// All blocking happens inside the hook; the dispatcher loop itself
-/// stays linear and tight.
-pub trait DebugHooks {
-    /// Called before executing instruction `pc` in `fn_name`.
-    ///
-    /// `depth` is the number of *parent* frames currently on the stack
-    /// (top-level main = 0, a function called from main = 1, …).
-    ///
-    /// `frame` exposes the local register state for variable
-    /// inspection.  Implementations that don't need locals may
-    /// ignore it.
-    fn before_instruction(
-        &mut self,
-        fn_name: &str,
-        depth: usize,
-        pc: usize,
-        frame: &FrameView<'_>,
-    );
-
-    /// Called once when `dispatch` is about to return from `fn_name`
-    /// (whether via `ret`, falling off the end, or propagated error).
-    ///
-    /// The default impl is a no-op; servers that track call-stack
-    /// state via depth deltas only generally don't override this.
-    fn on_function_exit(&mut self, _fn_name: &str, _depth: usize) {}
+impl<'a> vm_debug::DebugFrame for FrameView<'a> {
+    fn register_names(&self) -> Vec<String> {
+        FrameView::register_names(self)
+    }
+    fn read_register(&self, name: &str) -> Option<String> {
+        FrameView::read_register(self, name)
+    }
 }
 
 // ---------------------------------------------------------------------------
@@ -126,8 +108,9 @@ mod tests {
     use super::*;
     use std::sync::{Arc, Mutex};
 
-    /// Recording hook used by tests to assert `before_instruction` was
-    /// called with the expected sequence of `(fn, pc, depth)` tuples.
+    /// Recording hook used by tests to assert `before_instruction`
+    /// was called with the expected sequence of `(fn, pc, depth)`
+    /// tuples.
     #[derive(Default)]
     struct RecordingHook {
         events: Arc<Mutex<Vec<(String, usize, usize)>>>,
@@ -139,7 +122,7 @@ mod tests {
             fn_name: &str,
             depth: usize,
             pc: usize,
-            _frame: &FrameView<'_>,
+            _frame: &dyn DebugFrame,
         ) {
             self.events.lock().unwrap().push((fn_name.to_string(), depth, pc));
         }
