+++ code/packages/rust/vm-debug/BUILD
@@ -1,0 +1,1 @@
+cargo test -p vm-debug -- --nocapture
