+++ code/packages/rust/vm-debug/src/lib.rs
@@ -1,0 +1,751 @@
+//! # `vm-debug` — language-agnostic debug substrate for IIR-based VMs.
+//!
+//! Exposes the [`DebugFrame`] + [`DebugHooks`] traits, the
+//! wire-format-aligned [`StopReason`] enum, and the production
+//! TCP-backed [`DebugServer`] that every LANG-VM-based interpreter
+//! shares.  See [`README.md`](../README.md) for the motivation and
+//! plug-in story.
+//!
+//! ## Wire protocol
+//!
+//! Newline-delimited JSON over TCP.  Same shape on every VM:
+//!
+//! ```text
+//! → { "cmd": "set_breakpoint", "fn": "foo", "offset": 5 }
+//! ← { "ok": true }
+//! ⇐ { "event": "stopped", "reason": "breakpoint", "fn": "foo", "offset": 5 }
+//! ```
+//!
+//! See [`dap_adapter_core::vm_conn`](../../dap-adapter-core/src/vm_conn.rs)
+//! for the adapter-side counterpart.
+
+#![warn(missing_docs)]
+#![warn(rust_2018_idioms)]
+
+use std::collections::{HashMap, HashSet, VecDeque};
+use std::io::{BufRead, BufReader, Read, Write};
+use std::net::TcpStream;
+
+use serde_json::{json, Value};
+
+// ===========================================================================
+// DebugFrame — VM-agnostic read-only frame view
+// ===========================================================================
+
+/// Read-only view into a VM's currently-executing frame.
+///
+/// Per-language VMs (twig-vm, the upcoming basic-vm / nib-vm / oct-vm)
+/// implement this trait on their own `FrameView`-style wrapper so
+/// `DebugServer` can render `variables` and respond to `get_slot`
+/// without depending on any concrete VM's `Frame` type.
+///
+/// The trait deliberately exposes everything as `String` — debug
+/// representations, not typed values.  This keeps `vm-debug` free of
+/// any value-model crate (`lispy-runtime`, `interpreter-ir::Operand`,
+/// …) and lets each VM choose its own renderer.
+pub trait DebugFrame {
+    /// Names of every register currently bound in this frame.
+    ///
+    /// Returned in implementation-defined order; consumers that need
+    /// stable order should sort.  `DebugServer::snapshot_frame` does.
+    fn register_names(&self) -> Vec<String>;
+
+    /// Printable representation of `name`'s current value, or `None`
+    /// if the register is not bound in this frame.
+    fn read_register(&self, name: &str) -> Option<String>;
+}
+
+// ===========================================================================
+// DebugHooks — interface the dispatcher calls at every safepoint
+// ===========================================================================
+
+/// The interface a VM dispatcher calls at every safepoint.
+///
+/// Implementations are free to:
+///
+/// 1. Process incoming wire-protocol commands (set/clear breakpoint).
+/// 2. Detect that the current `(fn_name, pc)` is a breakpoint and emit
+///    a `stopped` event.
+/// 3. **Block** on the wire (e.g. waiting for `continue` after a stop)
+///    — the dispatcher resumes only when the hook returns.
+/// 4. Implement single-step by remembering "stop on the *next*
+///    instruction" between calls.
+///
+/// All blocking happens inside the hook; the dispatcher loop itself
+/// stays linear and tight.
+///
+/// The trait is object-safe so dispatchers can store
+/// `&mut dyn DebugHooks` cheaply.  The frame parameter is
+/// `&dyn DebugFrame` for the same reason.
+pub trait DebugHooks {
+    /// Called before executing instruction `pc` in `fn_name`.
+    ///
+    /// `depth` is the number of *parent* frames currently on the
+    /// stack (top-level main = 0, a function called from main = 1, …).
+    ///
+    /// `frame` exposes the local register state for variable
+    /// inspection.  Implementations that don't need locals may
+    /// ignore it.
+    fn before_instruction(
+        &mut self,
+        fn_name: &str,
+        depth: usize,
+        pc: usize,
+        frame: &dyn DebugFrame,
+    );
+
+    /// Called once when `dispatch` is about to return from `fn_name`
+    /// (whether via `ret`, falling off the end, or propagated error).
+    ///
+    /// The default impl is a no-op; servers that track call-stack
+    /// state via depth deltas only generally don't override this.
+    fn on_function_exit(&mut self, _fn_name: &str, _depth: usize) {}
+}
+
+// ===========================================================================
+// StopReason — wire-format-aligned reasons the VM is stopped
+// ===========================================================================
+
+/// Why the VM is stopped at a safepoint.
+///
+/// String forms (`"breakpoint"`, `"step"`, `"pause"`, `"entry"`)
+/// match what `dap-adapter-core` maps to DAP `stopped` reasons.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum StopReason {
+    /// User-set breakpoint at this `(fn, pc)`.
+    Breakpoint,
+    /// Single-step completed.
+    Step,
+    /// Async pause request.
+    Pause,
+    /// Initial entry stop (always sent right after the VM starts).
+    Entry,
+}
+
+impl StopReason {
+    /// Wire-format string for this reason.
+    pub fn as_str(self) -> &'static str {
+        match self {
+            StopReason::Breakpoint => "breakpoint",
+            StopReason::Step       => "step",
+            StopReason::Pause      => "pause",
+            StopReason::Entry      => "entry",
+        }
+    }
+}
+
+// ===========================================================================
+// DoS guard
+// ===========================================================================
+
+/// Maximum bytes a single newline-terminated command line may contain.
+///
+/// Commands and event payloads are small structured JSON; 1 MiB is
+/// far above any legitimate value and well below memory exhaustion.
+/// Without this cap, a peer that streams bytes without a newline
+/// would force unbounded `String` growth.
+pub const MAX_LINE_BYTES: usize = 1024 * 1024;
+
+/// Read one line of at most [`MAX_LINE_BYTES`] bytes (inclusive of
+/// `\n`) from a `BufRead`, returning the line as a `String`.
+///
+/// Behaves like `BufRead::read_line` but with a hard cap: if more
+/// than [`MAX_LINE_BYTES`] bytes are consumed before a newline,
+/// returns an `InvalidData` error and the partially-read bytes are
+/// discarded.
+fn read_line_capped(reader: &mut impl BufRead, line: &mut String) -> std::io::Result<usize> {
+    let mut limited = reader.take(MAX_LINE_BYTES as u64);
+    let n = limited.read_line(line)?;
+    if n == MAX_LINE_BYTES && !line.ends_with('\n') {
+        return Err(std::io::Error::new(
+            std::io::ErrorKind::InvalidData,
+            format!("vm-debug line exceeded {MAX_LINE_BYTES} bytes"),
+        ));
+    }
+    Ok(n)
+}
+
+// ===========================================================================
+// DebugServer — TCP-backed production hook
+// ===========================================================================
+
+/// One-connection-per-VM debug server.
+///
+/// Owns the TCP stream, breakpoint set, single-step flag, pause flag,
+/// and a snapshot of the live call stack.  Created by the per-VM CLI
+/// when `--debug-port` is supplied; passed to the per-VM dispatcher
+/// as the `&mut dyn DebugHooks`.
+pub struct DebugServer {
+    reader: BufReader<TcpStream>,
+    writer: TcpStream,
+    breakpoints: HashSet<(String, usize)>,
+    single_step: bool,
+    pause_requested: bool,
+    /// Reconstructed call stack — `[(fn_name, pc), …]`, depth-indexed.
+    /// Updated from `before_instruction`'s `(fn, depth, pc)` triple.
+    call_stack: Vec<(String, usize)>,
+    /// `true` once the initial-entry stop has been sent.
+    started: bool,
+    /// Last frame's register values, for `get_slot` queries.  Cleared
+    /// on every `before_instruction` and refilled from the
+    /// [`DebugFrame`].
+    last_frame_registers: Vec<(String, String)>,
+    /// Stable per-function slot assignment: `fn_name → sorted variable names`.
+    ///
+    /// Built from the IIR module at construction time (via
+    /// [`Self::new_with_module`]).  When this map contains an entry
+    /// for the current function, `snapshot_frame` uses it to assign
+    /// slot indices in alphabetical order over **all** variables of
+    /// the function — not just the ones that have been assigned so
+    /// far.  This produces stable, predictable slot indices that
+    /// match what per-language `build_sidecar` impls emit for
+    /// `reg_index`.
+    ///
+    /// When empty (plain `new` was used), `snapshot_frame` falls back
+    /// to the original behaviour: sort whatever variable names exist
+    /// in the live frame.
+    fn_sorted_vars: HashMap<String, Vec<String>>,
+}
+
+impl DebugServer {
+    /// Wrap a connected `TcpStream` as a debug server.
+    ///
+    /// The stream's `try_clone` must succeed (used to split read/write).
+    ///
+    /// Use [`Self::new_with_module`] instead when the IIR module is
+    /// available so that `get_slot` returns stable,
+    /// sidecar-compatible slot indices.
+    pub fn new(stream: TcpStream) -> std::io::Result<Self> {
+        let writer = stream.try_clone()?;
+        let reader = BufReader::new(stream);
+        Ok(DebugServer {
+            reader,
+            writer,
+            breakpoints:           HashSet::new(),
+            single_step:           false,
+            pause_requested:       false,
+            call_stack:            Vec::new(),
+            started:               false,
+            last_frame_registers:  Vec::new(),
+            fn_sorted_vars:        HashMap::new(),
+        })
+    }
+
+    /// Wrap a connected `TcpStream` as a debug server, pre-loading
+    /// stable slot indices from the IIR module.
+    ///
+    /// For each function in `module`, collects all variable names
+    /// (parameters and instruction `dest` fields) and sorts them
+    /// alphabetically.  The resulting list is the **stable slot
+    /// assignment** used by `get_slot`: slot `N` always corresponds
+    /// to the Nth alphabetically-sorted variable of the current
+    /// function, regardless of how many variables have been assigned
+    /// at the current instruction pointer.
+    ///
+    /// This matches the ordering per-language `build_sidecar` impls
+    /// emit for `reg_index`, so the DAP `variables` panel shows the
+    /// correct values.
+    pub fn new_with_module(
+        stream: TcpStream,
+        module: &interpreter_ir::IIRModule,
+    ) -> std::io::Result<Self> {
+        let mut srv = Self::new(stream)?;
+        for func in &module.functions {
+            let mut names: HashSet<String> = HashSet::new();
+            for (param_name, _) in &func.params {
+                names.insert(param_name.clone());
+            }
+            for instr in &func.instructions {
+                if let Some(dest) = &instr.dest {
+                    names.insert(dest.clone());
+                }
+            }
+            let mut sorted: Vec<String> = names.into_iter().collect();
+            sorted.sort();
+            srv.fn_sorted_vars.insert(func.name.clone(), sorted);
+        }
+        Ok(srv)
+    }
+
+    // -------------------------------------------------------------------
+    // Wire I/O
+    // -------------------------------------------------------------------
+
+    /// Send a single JSON line to the adapter.
+    fn write_json(&mut self, v: Value) -> std::io::Result<()> {
+        let mut s = serde_json::to_string(&v)?;
+        s.push('\n');
+        self.writer.write_all(s.as_bytes())?;
+        self.writer.flush()
+    }
+
+    /// Read the next JSON line, blocking.  Returns `None` on EOF.
+    ///
+    /// Uses the [`MAX_LINE_BYTES`]-capped reader so a malicious peer
+    /// cannot OOM us with a never-newline-terminated stream.
+    fn read_json_blocking(&mut self) -> std::io::Result<Option<Value>> {
+        let mut line = String::new();
+        if read_line_capped(&mut self.reader, &mut line)? == 0 {
+            return Ok(None); // EOF — adapter disconnected
+        }
+        let v: Value = serde_json::from_str(line.trim_end())
+            .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData,
+                                             format!("bad JSON: {e}")))?;
+        Ok(Some(v))
+    }
+
+    /// Drain *any number of pending* commands without blocking.
+    ///
+    /// Uses `set_nonblocking(true)` for the duration of the call.
+    /// Returns every pending command parsed as JSON; commands that
+    /// don't parse are silently dropped (the adapter sends only
+    /// well-formed JSON).
+    fn drain_pending(&mut self) -> std::io::Result<VecDeque<Value>> {
+        let mut out = VecDeque::new();
+        self.writer.set_nonblocking(true)?;
+        loop {
+            let mut line = String::new();
+            match read_line_capped(&mut self.reader, &mut line) {
+                Ok(0) => break,
+                Ok(_) => {
+                    if let Ok(v) = serde_json::from_str::<Value>(line.trim_end()) {
+                        out.push_back(v);
+                    }
+                }
+                Err(e) if e.kind() == std::io::ErrorKind::WouldBlock => break,
+                Err(e) => {
+                    self.writer.set_nonblocking(false)?;
+                    return Err(e);
+                }
+            }
+        }
+        self.writer.set_nonblocking(false)?;
+        Ok(out)
+    }
+
+    // -------------------------------------------------------------------
+    // Command handling
+    // -------------------------------------------------------------------
+
+    /// Process one command; return `true` if the command unblocks the
+    /// VM (i.e. `continue`, `step_instruction`).
+    ///
+    /// Synchronous commands (`set_breakpoint`, `pause`,
+    /// `get_call_stack`, `get_slot`) are handled inline and **do
+    /// not** unblock — caller keeps reading until a continue/step
+    /// arrives.
+    fn handle_command(&mut self, v: Value) -> std::io::Result<bool> {
+        let cmd = v.get("cmd").and_then(|c| c.as_str()).unwrap_or("");
+        match cmd {
+            "set_breakpoint" => {
+                if let (Some(f), Some(o)) = (v.get("fn").and_then(|x| x.as_str()),
+                                              v.get("offset").and_then(|x| x.as_u64())) {
+                    self.breakpoints.insert((f.to_string(), o as usize));
+                }
+                self.write_json(json!({"ok": true}))?;
+                Ok(false)
+            }
+            "clear_breakpoint" => {
+                if let (Some(f), Some(o)) = (v.get("fn").and_then(|x| x.as_str()),
+                                              v.get("offset").and_then(|x| x.as_u64())) {
+                    self.breakpoints.remove(&(f.to_string(), o as usize));
+                }
+                self.write_json(json!({"ok": true}))?;
+                Ok(false)
+            }
+            "continue" => {
+                self.single_step = false;
+                self.pause_requested = false;
+                self.write_json(json!({"ok": true}))?;
+                Ok(true)
+            }
+            "pause" => {
+                self.pause_requested = true;
+                self.write_json(json!({"ok": true}))?;
+                Ok(false)
+            }
+            "step_instruction" => {
+                self.single_step = true;
+                self.pause_requested = false;
+                self.write_json(json!({"ok": true}))?;
+                Ok(true)
+            }
+            "get_call_stack" => {
+                let frames: Vec<Value> = self.call_stack.iter()
+                    .rev() // deepest frame first per protocol
+                    .map(|(f, o)| json!({"fn": f, "offset": o}))
+                    .collect();
+                self.write_json(json!({"frames": frames}))?;
+                Ok(false)
+            }
+            "get_slot" => {
+                let slot = v.get("slot").and_then(|s| s.as_u64()).unwrap_or(0) as usize;
+                let repr = self.last_frame_registers.get(slot)
+                    .map(|(_, r)| r.clone())
+                    .unwrap_or_else(|| "<undef>".to_string());
+                self.write_json(json!({"kind": "any", "repr": repr}))?;
+                Ok(false)
+            }
+            other => {
+                self.write_json(json!({"ok": false, "error": format!("unknown cmd: {other}")}))?;
+                Ok(false)
+            }
+        }
+    }
+
+    // -------------------------------------------------------------------
+    // Stop coordination
+    // -------------------------------------------------------------------
+
+    fn emit_stopped(&mut self, reason: StopReason, fn_name: &str, pc: usize) -> std::io::Result<()> {
+        self.write_json(json!({
+            "event":  "stopped",
+            "reason": reason.as_str(),
+            "fn":     fn_name,
+            "offset": pc,
+        }))
+    }
+
+    /// Block reading commands until one returns `true` (continue /
+    /// step).
+    fn block_until_resume(&mut self, fn_name: &str, pc: usize) -> std::io::Result<()> {
+        let _ = (fn_name, pc); // captured implicitly in `emit_stopped` callers above
+        loop {
+            match self.read_json_blocking()? {
+                Some(v) => {
+                    if self.handle_command(v)? {
+                        return Ok(());
+                    }
+                }
+                None => {
+                    // Adapter disconnected; let dispatch continue freely.
+                    return Ok(());
+                }
+            }
+        }
+    }
+
+    // -------------------------------------------------------------------
+    // Stack-tracking
+    // -------------------------------------------------------------------
+
+    /// Update `call_stack` based on the new `(fn_name, depth, pc)` triple.
+    ///
+    /// `depth` is the number of *parent* frames so the array is sized
+    /// to `depth + 1` after this call.
+    fn track_stack(&mut self, fn_name: &str, depth: usize, pc: usize) {
+        self.call_stack.truncate(depth);
+        while self.call_stack.len() < depth {
+            self.call_stack.push(("<unknown>".to_string(), 0));
+        }
+        if self.call_stack.len() == depth {
+            self.call_stack.push((fn_name.to_string(), pc));
+        } else {
+            self.call_stack[depth] = (fn_name.to_string(), pc);
+        }
+    }
+
+    /// Snapshot the [`DebugFrame`]'s registers into
+    /// `last_frame_registers`.
+    ///
+    /// ## Slot-index strategy
+    ///
+    /// If a stable variable list was pre-loaded for `fn_name` (via
+    /// [`Self::new_with_module`]), `last_frame_registers[N]` is the
+    /// Nth alphabetically-sorted variable of the function — even if
+    /// that variable hasn't been assigned yet (it gets the string
+    /// `"<undef>"`).  This produces slot indices that are
+    /// **constant across all breakpoints** in the function and
+    /// match what the debug sidecar declares.
+    ///
+    /// Fallback (no module was loaded): sort the names that happen
+    /// to be in the live frame right now — the original behaviour.
+    fn snapshot_frame(&mut self, fn_name: &str, frame: &dyn DebugFrame) {
+        self.last_frame_registers.clear();
+        if let Some(sorted) = self.fn_sorted_vars.get(fn_name) {
+            for name in sorted {
+                let r = frame.read_register(name)
+                    .unwrap_or_else(|| "<undef>".to_string());
+                self.last_frame_registers.push((name.clone(), r));
+            }
+        } else {
+            let mut names = frame.register_names();
+            names.sort();
+            for n in names {
+                let r = frame.read_register(&n)
+                    .unwrap_or_else(|| "<undef>".to_string());
+                self.last_frame_registers.push((n, r));
+            }
+        }
+    }
+
+    // -------------------------------------------------------------------
+    // Public lifecycle (called by the CLI)
+    // -------------------------------------------------------------------
+
+    /// Send the initial "entry" stop and block until the adapter
+    /// replies with `continue` (or `step_instruction`).
+    ///
+    /// The CLI calls this once *before* invoking the dispatcher's
+    /// `run_with_debug` entry point so the adapter has a chance to
+    /// install breakpoints before any code runs.
+    pub fn await_initial_continue(&mut self) -> std::io::Result<()> {
+        self.emit_stopped(StopReason::Entry, "<entry>", 0)?;
+        self.started = true;
+        self.block_until_resume("<entry>", 0)
+    }
+
+    /// Send the final "exited" event when the VM finishes.
+    pub fn emit_exited(&mut self, exit_code: i32) -> std::io::Result<()> {
+        self.write_json(json!({
+            "event":     "exited",
+            "exit_code": exit_code,
+        }))
+    }
+}
+
+// ===========================================================================
+// DebugHooks impl
+// ===========================================================================
+
+impl DebugHooks for DebugServer {
+    fn before_instruction(
+        &mut self,
+        fn_name: &str,
+        depth: usize,
+        pc: usize,
+        frame: &dyn DebugFrame,
+    ) {
+        self.track_stack(fn_name, depth, pc);
+        self.snapshot_frame(fn_name, frame);
+
+        // 1. Drain any pending commands (set_breakpoint, etc.) that
+        //    arrived while the VM was running.
+        if let Ok(pending) = self.drain_pending() {
+            for cmd in pending {
+                let _ = self.handle_command(cmd);
+            }
+        }
+
+        // 2. Determine whether this safepoint warrants a stop.
+        let stop_reason = if self.breakpoints.contains(&(fn_name.to_string(), pc)) {
+            Some(StopReason::Breakpoint)
+        } else if self.single_step {
+            self.single_step = false;
+            Some(StopReason::Step)
+        } else if self.pause_requested {
+            self.pause_requested = false;
+            Some(StopReason::Pause)
+        } else {
+            None
+        };
+
+        if let Some(reason) = stop_reason {
+            let _ = self.emit_stopped(reason, fn_name, pc);
+            let _ = self.block_until_resume(fn_name, pc);
+        }
+    }
+}
+
+// ===========================================================================
+// Tests
+// ===========================================================================
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::net::TcpListener;
+    use std::sync::{Arc, Mutex};
+    use std::thread;
+
+    // ── DebugHooks trait — minimal recording impl ──────────────────────
+
+    /// Recording hook used by trait-shape tests.
+    #[derive(Default)]
+    struct RecordingHook {
+        events: Arc<Mutex<Vec<(String, usize, usize)>>>,
+    }
+
+    impl DebugHooks for RecordingHook {
+        fn before_instruction(
+            &mut self,
+            fn_name: &str,
+            depth: usize,
+            pc: usize,
+            _frame: &dyn DebugFrame,
+        ) {
+            self.events.lock().unwrap().push((fn_name.to_string(), depth, pc));
+        }
+    }
+
+    /// Concrete DebugFrame for trait tests.
+    struct VecFrame(Vec<(String, String)>);
+    impl DebugFrame for VecFrame {
+        fn register_names(&self) -> Vec<String> {
+            self.0.iter().map(|(n, _)| n.clone()).collect()
+        }
+        fn read_register(&self, name: &str) -> Option<String> {
+            self.0.iter().find(|(n, _)| n == name).map(|(_, v)| v.clone())
+        }
+    }
+
+    #[test]
+    fn recording_hook_default_compiles() {
+        let _h = RecordingHook::default();
+    }
+
+    #[test]
+    fn recording_hook_records_calls() {
+        let mut h = RecordingHook::default();
+        let f = VecFrame(vec![]);
+        h.before_instruction("main", 0, 0, &f);
+        h.before_instruction("main", 0, 1, &f);
+        h.before_instruction("foo",  1, 0, &f);
+        assert_eq!(h.events.lock().unwrap().len(), 3);
+    }
+
+    #[test]
+    fn stop_reason_wire_strings_stable() {
+        assert_eq!(StopReason::Breakpoint.as_str(), "breakpoint");
+        assert_eq!(StopReason::Step.as_str(),       "step");
+        assert_eq!(StopReason::Pause.as_str(),      "pause");
+        assert_eq!(StopReason::Entry.as_str(),      "entry");
+    }
+
+    // ── DebugServer — loopback tests ───────────────────────────────────
+
+    /// Spawn a TcpListener on a random localhost port and return
+    /// `(server-side stream, client-side stream)`.
+    fn loopback_pair() -> (TcpStream, TcpStream) {
+        let listener = TcpListener::bind("127.0.0.1:0").unwrap();
+        let addr = listener.local_addr().unwrap();
+        let client_handle = thread::spawn(move || TcpStream::connect(addr).unwrap());
+        let (server, _) = listener.accept().unwrap();
+        let client = client_handle.join().unwrap();
+        (server, client)
+    }
+
+    fn read_line(s: &mut TcpStream) -> String {
+        let mut buf = [0u8; 4096];
+        let n = std::io::Read::read(s, &mut buf).unwrap();
+        String::from_utf8_lossy(&buf[..n]).to_string()
+    }
+
+    #[test]
+    fn server_constructs_from_stream() {
+        let (server_side, _client) = loopback_pair();
+        let _ = DebugServer::new(server_side).expect("ok");
+    }
+
+    #[test]
+    fn set_breakpoint_acks() {
+        let (server_side, mut client) = loopback_pair();
+        let mut server = DebugServer::new(server_side).unwrap();
+        client.write_all(b"{\"cmd\":\"set_breakpoint\",\"fn\":\"main\",\"offset\":5}\n").unwrap();
+        let cmd = server.read_json_blocking().unwrap().unwrap();
+        let unblocks = server.handle_command(cmd).unwrap();
+        assert!(!unblocks, "set_breakpoint must not unblock");
+        assert_eq!(server.breakpoints.len(), 1);
+        assert!(server.breakpoints.contains(&("main".to_string(), 5)));
+        let resp = read_line(&mut client);
+        assert!(resp.contains("\"ok\":true"), "got: {resp}");
+    }
+
+    #[test]
+    fn continue_unblocks() {
+        let (server_side, mut client) = loopback_pair();
+        let mut server = DebugServer::new(server_side).unwrap();
+        client.write_all(b"{\"cmd\":\"continue\"}\n").unwrap();
+        let cmd = server.read_json_blocking().unwrap().unwrap();
+        let unblocks = server.handle_command(cmd).unwrap();
+        assert!(unblocks);
+        let _ = read_line(&mut client);
+    }
+
+    #[test]
+    fn step_instruction_unblocks_and_sets_flag() {
+        let (server_side, mut client) = loopback_pair();
+        let mut server = DebugServer::new(server_side).unwrap();
+        client.write_all(b"{\"cmd\":\"step_instruction\"}\n").unwrap();
+        let cmd = server.read_json_blocking().unwrap().unwrap();
+        let unblocks = server.handle_command(cmd).unwrap();
+        assert!(unblocks);
+        assert!(server.single_step);
+        let _ = read_line(&mut client);
+    }
+
+    #[test]
+    fn track_stack_pushes_and_pops() {
+        let (s, _c) = loopback_pair();
+        let mut srv = DebugServer::new(s).unwrap();
+        srv.track_stack("main", 0, 0);
+        srv.track_stack("main", 0, 1);
+        srv.track_stack("foo", 1, 0);
+        assert_eq!(srv.call_stack, vec![
+            ("main".to_string(), 1),
+            ("foo".to_string(),  0),
+        ]);
+        srv.track_stack("main", 0, 2);
+        assert_eq!(srv.call_stack, vec![("main".to_string(), 2)]);
+    }
+
+    #[test]
+    fn read_line_capped_rejects_overlong_input() {
+        use std::io::Cursor;
+        let bytes = vec![b'A'; MAX_LINE_BYTES + 1];
+        let mut reader = std::io::BufReader::new(Cursor::new(bytes));
+        let mut line = String::new();
+        let err = read_line_capped(&mut reader, &mut line).unwrap_err();
+        assert_eq!(err.kind(), std::io::ErrorKind::InvalidData);
+    }
+
+    #[test]
+    fn read_line_capped_accepts_legitimate_line() {
+        use std::io::Cursor;
+        let raw = b"{\"cmd\":\"continue\"}\n".to_vec();
+        let mut reader = std::io::BufReader::new(Cursor::new(raw));
+        let mut line = String::new();
+        let n = read_line_capped(&mut reader, &mut line).unwrap();
+        assert_eq!(n, line.len());
+        assert!(line.ends_with('\n'));
+    }
+
+    #[test]
+    fn get_call_stack_returns_deepest_first() {
+        let (server_side, mut client) = loopback_pair();
+        let mut server = DebugServer::new(server_side).unwrap();
+        server.track_stack("main", 0, 0);
+        server.track_stack("foo",  1, 5);
+        client.write_all(b"{\"cmd\":\"get_call_stack\"}\n").unwrap();
+        let cmd = server.read_json_blocking().unwrap().unwrap();
+        let _ = server.handle_command(cmd).unwrap();
+        let resp = read_line(&mut client);
+        let idx_foo = resp.find("\"foo\"").expect("foo present");
+        let idx_main = resp.find("\"main\"").expect("main present");
+        assert!(idx_foo < idx_main, "deepest first: {resp}");
+    }
+
+    #[test]
+    fn snapshot_frame_uses_stable_slot_ordering_when_module_loaded() {
+        // With a stable per-fn slot list, snapshot_frame should walk
+        // it in alphabetical order, returning <undef> for unbound
+        // names — independent of the live frame's iteration order.
+        let (s, _c) = loopback_pair();
+        let mut srv = DebugServer::new(s).unwrap();
+        srv.fn_sorted_vars.insert(
+            "main".to_string(),
+            vec!["a".to_string(), "b".to_string(), "c".to_string()],
+        );
+        let live = VecFrame(vec![
+            ("b".to_string(), "BBB".to_string()),
+            ("a".to_string(), "AAA".to_string()),
+            // "c" is missing → <undef>
+        ]);
+        srv.snapshot_frame("main", &live);
+        assert_eq!(srv.last_frame_registers, vec![
+            ("a".to_string(), "AAA".to_string()),
+            ("b".to_string(), "BBB".to_string()),
+            ("c".to_string(), "<undef>".to_string()),
+        ]);
+    }
+}
